+++ epidemic/__init__.py
@@ -1,1 +1,1 @@
-from epidemic.source import linear_regression, graph, Predict, Predict_Epidemic, Predict_Virus_Growth, train, cli
+from epidemic.source import linear_regression, graph, Predict, Predict_Epidemic, Predict_Virus_Growth, train, example